+++ PAN/src/main/java/com/pan/spring/dao/impl/JobSeekerDaoImpl.java
@@ -159,7 +159,7 @@ public class JobSeekerDaoImpl implements JobSeekerDao {
 		List<Integer> list = new ArrayList<Integer>();
 		List<?> querylist = query.getResultList();
 		for (Iterator<?> iterator = querylist.iterator(); iterator.hasNext();) {
-			int uid = (int) iterator.next();
+			int uid = (int)(Integer) iterator.next();
 			list.add(uid);
 		}
 		return list;
